+++ .circleci/config.yml
@@ -9,7 +9,7 @@ jobs:
       - checkout
       - node/with-cache:
           steps:
-            - run: yarn install
+            - run: pnpm install
             - run: npm test
             - run: npx semantic-release
 workflows:
