+++ package.json
@@ -11,8 +11,7 @@
     "index.js"
   ],
   "scripts": {
-    "test": "jest && npm run lint",
-    "lint": "xo index.js __test__/*.test.js"
+    "test": "node test/test.js"
   },
   "author": "egoist <0x142857@gmail.com>",
   "license": "MIT",
@@ -20,18 +19,10 @@
     "testEnvironment": "node"
   },
   "devDependencies": {
-    "eslint-config-rem": "^3.2.0",
-    "jest-cli": "^18.1.0",
-    "xo": "^0.17.1"
-  },
-  "xo": {
-    "extends": "rem/prettier",
-    "envs": [
-      "jest"
-    ],
-    "ignores": [
-      "**/*/fixture*.js"
-    ]
+    "@egoist/prettier-config": "^0.1.0",
+    "esm": "^3.2.25",
+    "prettier": "^2.0.5",
+    "uvu": "^0.2.2"
   },
   "dependencies": {
     "@babel/parser": "^7.10.5",
